블랙잭 관련 기능 구현 및 리팩토링 완료.

+++ src/main/java/com/github/callmewaggs/game/IOHelper.java
@@ -88,7 +88,7 @@ public class IOHelper {
     printMessage("\n## 최종 수익");
   }
 
-  public static void printNameAndIncome(Map<String, Long> participantsIncome) {
+  public static void printNameAndIncome(Map<String, Double> participantsIncome) {
     participantsIncome.forEach((name, income) -> printMessage(name + " : " + income));
   }
 }
